+++ Sprint-3/slideshow/includes/images.js
@@ -1,0 +1,6 @@
+export const images = [
+  "./assets/cute-cat-a.webp",
+  "./assets/cute-cat-b.webp",
+  "./assets/cute-cat-c.webp",
+  "./assets/cute-cat-d.webp",
+];
