+++ Sprint-3/slideshow/package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "license": "CC-BY-SA-4.0",
   "description": "You must update this package",
+  "type": "module",
   "scripts": {
     "test": "jest --config=../jest.config.js slideshow"
   },
